#45 - update and refactor checkout page
