+++ src/pages/checkout/information.jsx
@@ -1,0 +1,5 @@
+function CheckoutInformation() {
+  return <div>This is checkout information</div>
+}
+
+export default CheckoutInformation
