+++ src/pages/checkout/payment.jsx
@@ -1,0 +1,5 @@
+function CheckoutPayment() {
+  return <div>this is checkout payment</div>
+}
+
+export default CheckoutPayment
