+++ src/pages/checkout/shipping.jsx
@@ -1,0 +1,5 @@
+function CheckoutShipping() {
+  return <div>this is checkout shipping</div>
+}
+
+export default CheckoutShipping
